refactor: update styled-react imports and build scripts
