+++ packages/react/script/build
@@ -5,9 +5,6 @@ set -e
 # Clean up
 npm run clean
 
-# Generate color schemes
-npm run build:precompile-color-schemes
-
 # Generate types for CSS Modules
 npm run type-css-modules
