+++ packages/styled-react/package.json
@@ -22,9 +22,19 @@
   ],
   "scripts": {
     "build": "script/build",
+    "build:precompile-color-schemes": "tsx script/precompile-color-schemes.ts",
     "clean": "rimraf dist",
     "lint:npm": "publint --types",
     "type-check": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@styled-system/css": "^5.1.5",
+    "@styled-system/props": "^5.1.5",
+    "@styled-system/theme-get": "^5.1.2",
+    "@types/styled-system": "^5.1.23",
+    "@types/styled-system__css": "^5.0.16",
+    "@types/styled-system__theme-get": "^5.0.1",
+    "styled-system": "^5.1.5"
   },
   "devDependencies": {
     "@babel/preset-react": "^7.27.1",
@@ -33,7 +43,9 @@
     "@rollup/plugin-babel": "^6.0.4",
     "@types/react": "18.3.11",
     "@types/react-dom": "18.3.1",
+    "@types/styled-components": "^5.1.11",
     "@vitejs/plugin-react": "^4.3.3",
+    "babel-plugin-styled-components": "2.1.4",
     "publint": "^0.3.12",
     "react": "18.3.1",
     "react-dom": "18.3.1",
@@ -41,6 +53,7 @@
     "rollup": "4.50.1",
     "rollup-plugin-typescript2": "^0.36.0",
     "styled-components": "5.3.11",
+    "tsx": "4.20.3",
     "typescript": "^5.9.2"
   },
   "peerDependencies": {
